Promise 	if elseif else 	match as 	OK
